refactor:#99 getUserSessionState 함수 서버액션-> 클라이언트 / 서버용 로직으로 분리하여 이동
